Fix regex validator always matching whole message
This fixes the issue, that the reges validator always matches
the complete commit message, even when the custom regex
does not start with `^` and end with `$`.

For this not to lead to unexpected behaviour, we update
existing configurations in the way that we add `^` and `$`
when the original pattern does not contain them.

Committed-by: Eduard Heimbuch <eduard.heimbuch@cloudogu.com>
Co-authored-by: René Pfeuffer <rene.pfeuffer@cloudogu.com>